Update the readme for new grant usage
The work in the repo is all built on top of
https://github.com/LBHackney-IT/discretionary-business-grants and as
both applications are likely to run at the same time, this was the
simplest way to get modified functionality within a very short time frame.

So this commit updates the README to reflect that while including all
relevant documentation.