Update Validations to match expected data
The following validations needed updating, as the data might not fit the data provided:

`contactTypeId` - this comes in the formData as a string

`businessCustomCategory` is only populated if "Other" is selected in the previous question.
A conditional validation should be used here, which will be added if time allows.

`businessRatesAccountNumber` can end in an 'X', so we can't rely on integer here. Again,
a conditional validation should be used here, which will be added if time allows.

`businessWebsite` shouldn't be required (market stalls are unlikely to have a website)