Update Application states values
This grant has a different processing flow, so this change updates the
application state values to reflect that flow.

The data has been changed in the migrations as that is where the initial
set up of application state was done, seeding with these values would
complicate this and affect migrations done in the future.

The migrate up (DELETE) won't work if any associations exist on the ids
above 9 exist, but as this is an initial set up I thought it would be OK
to have that in here.