Mutate formData Object to include businessIdentifyNumber
Adding the validation on types of company number (using different inputs
for each type) meant that the form validation on the back end failed.

To to satisfy both from and back end validations, I'm looping through
all the possible company numbers and when the property existis in the
form data I assign that to `business.businessIdentifyNumber` in the formData
and pass that to satisfy the back end.

Not very happy with this fix, a better way of handling this is surely
possible but I can't think of it right now.