Remove truthy value check on admin form
The filter on this code means that only truthy values would be displayed
on the admin check form. However the admin people need to verify that
served_legal_notices checkbox is False, so we should just show everything
to resolve that. Only a few address fields and webiste address fields
will be shown to make the page longer, but this is a quick fix.