Revert d4e46fe9 - add in check for truthy values
This might be breaking viewing some applications due to trying to coerce
null values in some fields, depending on what users submit. Hopefully
this allows apploications to be viewed and we will have to use a more
complex filter to check for the individual key on `servedLegalNotices`
in the application data.