Remove unused updateApplication usecase
Not used in the current iteration of the system and was not initialised
correctly in the AppContainer.

Tests were sporadically failing due to promises not resolving correctly.
You'll notice the signature of the usecase:
```javascript
export const updateApplication = async ({ clientGeneratedId, data, user })
```
Does not match the invocation from the AppController:
```javascript
return updateApplication(this);
```
Where `this` is the AppContainer not an object with clientGeneratedId,
data, or user properties.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch master
# Your branch is up-to-date with 'origin/master'.
#
# Changes to be committed:
#	modified:   containers/AppContainer.js
#	modified:   containers/AppContainer.test.js
#	modified:   lib/usecases/updateApplication.js
#
# Changes not staged for commit:
#	modified:   .env.sample
#	modified:   README.md
#	modified:   utils/fileUpload.js
#